+++ src/ts-framework/Core/Configuration.ts
@@ -1,4 +1,4 @@
-///<reference path="../../typings/main.d.ts" />
+///<reference path="../../../typings/main.d.ts" />
 
 import * as _ from "lodash";
