+++ src/ts-framework/Core/Cookie.ts
