+++ src/ts-framework/Core/Exception/Exception.ts
@@ -1,10 +1,4 @@
-export declare class Error
-{
-    public name: string;
-    public message: string;
-    public stack: string;
-    constructor(message?: string);
-}
+import {Error} from "./Error";
 
 export class Exception extends Error
 {
