+++ src/ts-framework/Http/IMultipleResultCallback.ts
@@ -1,0 +1,3 @@
+/**
+ * Created by jhon on 27/04/16.
+ */
