+++ src/ts-framework/Model/DatabaseDriver/DatabaseDriverEnd.ts
@@ -1,0 +1,17 @@
+/**
+ * TS-Framework DatabaseDriver
+ * Contains method for oparation with WL Query
+ * @parameter T
+ */
+
+//needs to refactor class name !!!
+export class DatabaseDriverEnd<T> implements DatabaseDriverEnd<T> {
+
+    /**
+     * Register query method WHERE
+     * @param {IMultipleResultCallback<T>} callback
+     */
+    public done(callback: IMultipleResultCallback<T>) {
+
+    }
+}
