+++ src/ts-framework/Model/DatabaseDriver/DatabaseDriverRaw.ts
@@ -1,0 +1,17 @@
+/**
+ * TS-Framework DatabaseDriverUnique
+ * Contains method for oparation with WL Query
+ * @parameter T
+ */
+
+//needs to refactor class name !!!
+export class DatabaseDriverRaw<T> implements IDatabaseDriverRaw<T> {
+
+    /**
+     * Register query method WHERE
+     * @param {ISingleResultCallback<T>} callback
+     */
+    public done(callback: ISingleResultCallback<T>) {
+
+    }
+}
