+++ src/ts-framework/Model/DatabaseDriver/DatabaseDriverUnique.ts
@@ -1,0 +1,17 @@
+/**
+ * TS-Framework DatabaseDriverUnique
+ * Contains method for oparation with WL Query
+ * @parameter T
+ */
+
+//needs to refactor class name !!!
+export class DatabaseDriverUnique<T> implements DatabaseDriverUnique<T> {
+
+    /**
+     * Register query method WHERE
+     * @param {ISingleResultCallback<T>} callback
+     */
+    public done(callback: ISingleResultCallback<T>) {
+
+    }
+}
