+++ src/ts-framework/Model/DatabaseDriver/IDatabaseDriverEnd.ts
@@ -1,0 +1,8 @@
+export interface IDatabaseDriverEnd<T> {
+
+    /**
+     * Register query method WHERE
+     * @param {IMultipleResultCallback<T>} callback
+     */
+    done(callback: IMultipleResultCallback<T>);
+}
