+++ src/ts-framework/Model/DatabaseDriver/IDatabaseDriverRaw.ts
@@ -1,0 +1,8 @@
+export interface IDatabaseDriverRaw<T> {
+
+    /**
+     * Register query method WHERE
+     * @param {ISingleResultCallback<T>} callback
+     */
+    done(callback: ISingleResultCallback<T>);
+}
