+++ src/ts-framework/Model/DatabaseDriver/IDatabaseDriverUnique.ts
@@ -1,0 +1,8 @@
+export interface IDatabaseDriverUnique<T> {
+
+    /**
+     * Register query method WHERE
+     * @param {ISingleResultCallback<T>} callback
+     */
+    done(callback: ISingleResultCallback<T>);
+}
