+++ src/ts-framework/Router/RouteCollection.ts
@@ -1,0 +1,3 @@
+/**
+ * Created by jhon on 27/04/16.
+ */
