--- src/ts-framework/TS-Framework.ts
@@ -1,9 +1,0 @@
-export * from "./ActionFilter";
-export * from "./ActionResult";
-export * from "./Application";
-export * from "./Configuration";
-export * from "./Controller";
-export * from "./DatabaseDriver";
-export * from "./Http";
-export * from "./ModelValidation";
-export * from "./Router";
